+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-  - [AWS Route53 provider] Support customization of DNS record TTL through the use of annotation `external-dns.alpha.kubernetes.io/ttl` on services or ingresses
+  - [AWS Route53 provider] Support customization of DNS record TTL through the use of annotation `external-dns.alpha.kubernetes.io/ttl` on services or ingresses (#320) @kevinjqiu
+  - Added support for [DNSimple](https://dnsimple.com/) as DNS provider (#224) @jose5918
 
 ## v0.4.5 - 2017-09-24